requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
opencv-python==4.8.0.76
ultralytics==8.3.28
numpy==1.24.3
python-socketio==5.9.0
eventlet==0.33.3
pymongo==4.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0 